change name of mixin to prevent name similarity
Move the `observedAttributes` initialization into the mixin
